requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
distro==1.9.0
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
openai==1.41.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
tqdm==4.66.5
typing_extensions==4.12.2
uvicorn==0.30.5
